PR comments, move contain document search to ContainedDocument class.
